Switch to use the new etd_embargo_date method
The get_embargo_date was deprecated and then removed, but since there was no
test coverage, we never saw the deprecation
